windowheader in a new control
