+++ MainWindow.xaml.cs
@@ -52,8 +52,7 @@ namespace shufflecad_4
                 },
             };
 
-            LanguageCB.ItemsSource = languageClasses;
-            LanguageCB.SelectedIndex = 0;
+            ThisHeader.InitLangCB(languageClasses);
 
             connectionPage = new ConnectionPage();
             runPage = new RunPage();
@@ -74,8 +73,7 @@ namespace shufflecad_4
 
             loggerPage.AppendTextToLog("Shufflecad inited", "#00AEAE");
 
-            ProgressAsyncPB.Visibility = Visibility.Collapsed;
-            AllDonePI.Visibility = Visibility.Visible;
+            ThisHeader.InitDone();
         }
 
         public const string STATE_DEFAULT_COLOR = "#F0F8FF";
@@ -84,44 +82,21 @@ namespace shufflecad_4
 
         public void ChangeStateText(string text, string color = STATE_DEFAULT_COLOR)
         {
-            CurrentState.Text = text;
-            CurrentState.Foreground = new SolidColorBrush((Color)ColorConverter.ConvertFromString(color));
-        }
-
-        private void MinimizeButtonClick(object sender, RoutedEventArgs e)
-        {
-            this.WindowState = WindowState.Minimized;
-        }
-
-        private void MaximizeButton_Click(object sender, RoutedEventArgs e)
-        {
-            this.WindowState = WindowState.Maximized;
-        }
-
-        private void RestoreButton_Click(object sender, RoutedEventArgs e)
-        {
-            this.WindowState = WindowState.Normal;
-        }
-
-        async private void CloseButton_Click(object sender, RoutedEventArgs e)
-        {
-            await CommandsHelper.StopChecks();
-            await ConnectionHelper.StopHelper();
-            await JoystickHelper.StopJoystick();
-            this.Close();
+            ThisHeader.CurrentState.Text = text;
+            ThisHeader.CurrentState.Foreground = new SolidColorBrush((Color)ColorConverter.ConvertFromString(color));
         }
 
         private void RefreshMaximizeRestoreButton()
         {
             if (this.WindowState == WindowState.Maximized)
             {
-                this.maximizeButton.Visibility = Visibility.Collapsed;
-                this.restoreButton.Visibility = Visibility.Visible;
+                ThisHeader.maximizeButton.Visibility = Visibility.Collapsed;
+                ThisHeader.restoreButton.Visibility = Visibility.Visible;
             }
             else
             {
-                this.maximizeButton.Visibility = Visibility.Visible;
-                this.restoreButton.Visibility = Visibility.Collapsed;
+                ThisHeader.maximizeButton.Visibility = Visibility.Visible;
+                ThisHeader.restoreButton.Visibility = Visibility.Collapsed;
             }
         }
